+++ Packages/com.frameworkbase.core/UI/RedDotBadge.cs
@@ -1,5 +1,4 @@
 using System;
-using Framework.Foundation;
 using UnityEngine;
 using UnityEngine.Serialization;